# Conflicts:
#	discord_client_gateway/src/gateway.rs

+++ discord_client_gateway/src/events/structs/ready.rs
@@ -86,5 +86,5 @@ pub enum AuthenticatorType {
     Webauthn = 1,
     Totp = 2,
     Sms = 3,
-    Unknown(u8),
+    Unknown(u16),
 }
